requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-telegram-bot==21.3
google-generativeai==0.7.2
mysql-connector-python==9.0.0
httpx==0.27.0
python-dotenv==1.0.1